# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a api server for a shared task management tool",
        "title": "go-todo",
        "contact": {
            "name": "Ruizhe Zhang",
            "url": "https://github.com/nolanzzz"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/Users/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "User login",
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"data\":{\"token\":string},\"msg\":\"user login succeed\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/Users/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Register new account",
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"data\":{},\"msg\":\"user register succeed\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/ranking/minutes/:limit": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ranking"
                ],
                "summary": "Get ranking by total sum of minutes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "number of records",
                        "name": "limit",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"data\":{\"ranking\":{}},\"msg\":\"succeed\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/ranking/todos/:limit": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ranking"
                ],
                "summary": "Get ranking by number of todos",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "number of records",
                        "name": "limit",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"data\":{\"ranking\":{}},\"msg\":\"succeed\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/todo": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todo"
                ],
                "summary": "Get all todo tasks",
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"data\":{\"items\":{}},\"msg\":\"succeed\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todo"
                ],
                "summary": "Update an existing todo task",
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"data\":{},\"msg\":\"todo update succeed\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todo"
                ],
                "summary": "Create new todo task",
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"data\":{},\"msg\":\"todo create succeed\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/todo/:id": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todo"
                ],
                "summary": "Get one todo task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of task",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"data\":{\"item\":{}},\"msg\":\"succeed\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/todo/by/:userID": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todo"
                ],
                "summary": "Get all todo tasks of a specific user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of user",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"data\":{\"items\":{}},\"msg\":\"succeed\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/todo/done/:id": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todo"
                ],
                "summary": "Mark a task as completed",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of task",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"data\":{},\"msg\":\"todo completed\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/todo/undone/:id": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todo"
                ],
                "summary": "Undone a todo task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of task",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"data\":{},\"msg\":\"todo undone\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}